+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
-## [NEXT_RELEASE]
+## [NEXT_RELEASE] ## [0.3.0] - 2017-04-07
 ### Changed
 - `.teresaignore` pattern type from regex to glob